# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Source API",
        "description": "API for Source - Figma Design Audit Platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/projects/{projectId}/email-templates": {
            "get": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Get project email templates",
                "description": "Get all email templates for a project",
                "operationId": "27e4d8b77bc60443529bdccdd9aa82da",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of email templates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EmailTemplate"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/projects/{projectId}/email-templates/{emailTemplateId}": {
            "get": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Get email template by ID",
                "description": "Get a specific email template by its ID",
                "operationId": "4658b1d3952aec2ee0af7be093b94279",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "emailTemplateId",
                        "in": "path",
                        "description": "Email template ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email template data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailTemplate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Email template not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Update email template",
                "description": "Update an email template by its ID",
                "operationId": "ab091b63eac394fbe5c1071fc92bac58",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "emailTemplateId",
                        "in": "path",
                        "description": "Email template ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "section_name": {
                                        "type": "string",
                                        "example": "Updated Template Name",
                                        "nullable": true
                                    },
                                    "html": {
                                        "type": "string",
                                        "example": "<html>Updated content</html>",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email template updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailTemplate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Email template not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Delete email template",
                "description": "Delete an email template by its ID",
                "operationId": "8b41f1f4e4a472e67b580c93c5811e0a",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "emailTemplateId",
                        "in": "path",
                        "description": "Email template ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email template deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailTemplate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Email template not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/projects/{projectId}/email-templates/import-brevo": {
            "post": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Import Brevo template",
                "description": "Import an email template from Brevo",
                "operationId": "58eeb58931209fa44d74e2f1fe83eb90",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "brevo_template_id": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Template imported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailTemplate"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brevo API token not configured"
                    },
                    "404": {
                        "description": "Template not found in Brevo"
                    },
                    "409": {
                        "description": "Template already imported"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmailTemplate": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "section_name": {
                        "type": "string",
                        "example": "Welcome Email"
                    },
                    "campaign_id": {
                        "type": "integer",
                        "example": 123,
                        "nullable": true
                    },
                    "brevo_template_id": {
                        "type": "integer",
                        "example": 456,
                        "nullable": true
                    },
                    "html_content": {
                        "type": "string",
                        "example": "<html>...</html>",
                        "nullable": true
                    },
                    "thumbnail_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/thumb.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}